wip: refactor test registry helpers
